Update installation notes in README
I was reconnecting my Docker auto builds in light of the recent Docker
Hub hack, and realized that it's completely stupid and unacceptable for
me to be *recommending* bind-mounting the Docker socket into an svb
container without any discussion of the risks involved.

To be clear, I'm not *un*-recommending this strategy either, and I
absolutely still plan to distribute and use the svb Docker image. There
are a number of ways to both shoot yourself in the foot and apply proper
risk mitigation when getting software onto your systems. This is just
about having a more nuanced discussion of the risks and benefits for
particular use cases.

I've also added a bit more detail about manual installation behavior,
just to be clear that this is cool too.